* upstream/master: (100 commits)
  BibLaTeX mode: Add month and year field to deprecated tab (#2236)
  Change log level to debug
  Result of `scripts\syncLang.py update -e` and manual fixes
  Localization in Russian
  Update Menu_es.properties
  Update JabRef_es.properties
  Updated JabRef_tr_properties
  Update Chinese translation for version 3.7.
  Add issn field to drop-down (#2245)
  fix german backup related translations
  remove wrong translations
  Localization: JabRef_fr
  Fix ISBN #2244
  Update dependencies
  Global search usability improvements (#2234)
  Fixed problem where closing brackets could not be used as texts in la… (#1905)
  Disable auto complete for crossref field #2243
  Complete german translation (#2240)
  Fix infinite save as loop
  Fix NPE for filePath
  ...

+++ src/main/java/net/sf/jabref/logic/autosaveandbackup/BackupManager.java
@@ -5,7 +5,6 @@ import java.io.IOException;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.util.HashSet;
 import java.util.Optional;
@@ -21,6 +20,7 @@ import net.sf.jabref.logic.exporter.BibtexDatabaseWriter;
 import net.sf.jabref.logic.exporter.FileSaveSession;
 import net.sf.jabref.logic.exporter.SaveException;
 import net.sf.jabref.logic.exporter.SavePreferences;
+import net.sf.jabref.logic.util.io.FileUtil;
 import net.sf.jabref.model.database.BibDatabaseContext;
 import net.sf.jabref.model.database.event.BibDatabaseContextChangedEvent;
 import net.sf.jabref.preferences.JabRefPreferences;
@@ -32,12 +32,13 @@ import org.apache.commons.logging.LogFactory;
 /**
  * Backups the given bib database file from {@link BibDatabaseContext} on every {@link BibDatabaseContextChangedEvent}.
  * An intelligent {@link ExecutorService} with a {@link BlockingQueue} prevents a high load while making backups and rejects all redundant backup tasks.
+ * This class does not manage the .bak file which is created when opening a database.
  */
 public class BackupManager {
 
     private static final Log LOGGER = LogFactory.getLog(BackupManager.class);
 
-    private static final String BACKUP_FILENAME_ENDING = ".swp";
+    private static final String BACKUP_EXTENSION = ".sav";
 
     private static Set<BackupManager> runningInstances = new HashSet<>();
 
@@ -74,6 +75,10 @@ public class BackupManager {
 
     @Subscribe
     public synchronized void listen(@SuppressWarnings("unused") BibDatabaseContextChangedEvent event) {
+        startBackupTask();
+    }
+
+    private void startBackupTask() {
         try {
             executor.submit(backupTask);
         } catch (RejectedExecutionException e) {
@@ -84,18 +89,22 @@ public class BackupManager {
     /**
      * Unregisters the BackupManager from the eventBus of {@link BibDatabaseContext} and deletes the backup file.
      * This method should only be used when closing a database/JabRef legally.
-     *
-     * @param bibDatabaseContext Associated {@link BibDatabaseContext}
      */
     private void shutdown() {
         bibDatabaseContext.getDatabase().unregisterListener(this);
         bibDatabaseContext.getMetaData().unregisterListener(this);
         executor.shutdown();
         try {
-            Files.delete(backupPath);
+            if (Files.exists(backupPath) && !Files.isDirectory(backupPath)) {
+                Files.delete(backupPath);
+            }
         } catch (IOException e) {
             LOGGER.error("Error while deleting the backup file.", e);
         }
+    }
+
+    static Path getBackupPath(Path originalPath) {
+        return FileUtil.addExtension(originalPath, BACKUP_EXTENSION);
     }
 
     /**
@@ -111,7 +120,8 @@ public class BackupManager {
 
         if (originalFile.isPresent()) {
             backupManager.originalPath = originalFile.get().toPath();
-            backupManager.backupPath = Paths.get(backupManager.originalPath.toString() + BACKUP_FILENAME_ENDING);
+            backupManager.backupPath = getBackupPath(backupManager.originalPath);
+            backupManager.startBackupTask();
             bibDatabaseContext.getDatabase().registerListener(backupManager);
             bibDatabaseContext.getMetaData().registerListener(backupManager);
             runningInstances.add(backupManager);
@@ -139,7 +149,7 @@ public class BackupManager {
      * @param originalPath Path to the file a backup should be checked for.
      */
     public static boolean checkForBackupFile(Path originalPath) {
-        Path backupPath = Paths.get(originalPath.toString() + BACKUP_FILENAME_ENDING);
+        Path backupPath = getBackupPath(originalPath);
         return Files.exists(backupPath) && !Files.isDirectory(backupPath);
     }
 
@@ -149,7 +159,7 @@ public class BackupManager {
      * @param originalPath Path to the file which should be equalized to the backup file.
      */
     public static void restoreBackup(Path originalPath) {
-        Path backupPath = Paths.get(originalPath.toString() + BACKUP_FILENAME_ENDING);
+        Path backupPath = getBackupPath(originalPath);
         try {
             Files.copy(backupPath, originalPath, StandardCopyOption.REPLACE_EXISTING);
         } catch (IOException e) {
